% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikeDTGIW.R
\name{loglikeDTGIW}
\alias{loglikeDTGIW}
\title{Negative Log-Likelihood value of DTGIW distribution.}
\usage{
loglikeDTGIW(x, alpha, beta, lambda, theta)
}
\arguments{
\item{x}{a vector of quantile}

\item{alpha}{shape parameter#1}

\item{beta}{scale parameter}

\item{lambda}{shape pameter#2}

\item{theta}{the transmuted parameter}
}
\value{
the negative log-likelihood value of DTGIW distribution
}
\description{
The function for calculating negative log-likelihood value of DTGIW distribution.
}
\examples{
x <- rDTGIW(n=20,3.45,0.7,1.05,0)
loglikeDTGIW(x,3.45,0.7,1.05,0)
}
\references{
Atchanut Rattanalertnusorn and Sirinapa Aryuyuen (2021).
The zero-truncated discrete transmuted generalized inverse Weibull distribution and its applications,
Songklanakarin Journal of Science and Technology (SJST), Volume 43 No.4 (July - August 2021), pp. 1140 - 1151 <DOI: 10.14456/sjst-psu.2021.149>.
}
