% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDTGIW.R
\name{plotDTGIW}
\alias{plotDTGIW}
\title{Plot Discrete Transmuted Generalized Inverse Weibull(DTGIW) distribution.}
\usage{
plotDTGIW(x, fx, alpha = 3.45, beta = 0.7, lambda = 1.05, theta = 0)
}
\arguments{
\item{x}{a vector of quantile}

\item{fx}{probability mass function}

\item{alpha}{shape parameter#1.}

\item{beta}{scale parameter.}

\item{lambda}{shape pameter#2.}

\item{theta}{the transmuted parameter.}
}
\value{
the figure of DTGIW distribution
}
\description{
This function for the plot of DTGIW distribution.
}
\examples{
x <- c(0:10)
fx<- dDTGIW(x,3.45,0.7,1.05,0)
plotDTGIW(x,fx,alpha=3.45,beta=0.7,lambda=1.05,theta=0)

}
